+++ sharding-core/sharding-core-execute/src/test/resources/logback-test.xml
@@ -25,7 +25,7 @@
     <logger name="org.apache.shardingsphere" level="info" additivity="false">
         <appender-ref ref="console"/>
     </logger>
-    <logger name="org.apache.shardingsphere.sharding.execute.sql.execute.threadlocal.ExecutorExceptionHandler" level="off" />
+    <logger name="org.apache.shardingsphere.underlying.executor.sql.executor.ExecutorExceptionHandler" level="off" />
     <logger name="org.apache.shardingsphere.core.log.ConfigurationLogger" level="off" />
     <logger name="ShardingSphere-metadata" level="off" />
