+++ shardingsphere-underlying/shardingsphere-executor/src/main/java/org/apache/shardingsphere/underlying/executor/sql/connection/ConnectionMode.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.executor.sql.constant;
+package org.apache.shardingsphere.underlying.executor.sql.connection;
 
 /**
  * Connection Mode.
